requirements.txt
lxml==4.9.3
requests==2.31.0
cssselect==1.2.0
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
argparse==1.4.0
